fix: issues

+++ bot_commands/src/roll/dice.rs
@@ -18,7 +18,7 @@ impl Dice {
 impl RollDice for Dice {
     fn roll(&self) -> i64 {
         let mut rng = rand::rng();
-        rng.random_range(1..=self.sides) as i64
+        rng.random_range(1..=self.sides)
     }
 }
